requirements.txt
tkinter
opencv-python
Pillow
pygame
sounddevice
scipy
google-generativeai
speechrecognition
gtts
requests
